requirements.txt
django
django_rest_framework
djangorestframework-simplejwt
requests
django-cors-headers
django-bcrypt
pillow